requirements.txt
Django==4.2.7
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-ckeditor==6.6.0
Pillow==12.0.0